organized files into modules, added authenticate module
