+++ Components/Main/Lyrics.py
@@ -1,0 +1,10 @@
+from Components.Templates.TextLines import TextLines
+from Utils.lyrics import getLyrics
+
+
+class Lyrics(TextLines.TextLines):
+
+    def __init__(self, song, artist):
+        lyricLines = getLyrics(song, artist)
+        super().__init__("Lyrics", lyricLines)
+
