+++ clify
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from ViewManager import viewManager
-from Player import player
+import Components.Main.Player.player as player
 from Log import log
 from TitleBar import TitleBar
 from MainMenu import MainMenu
